Split income variables into separate files
